LayoutTests/inspector/debugger/linkifier-expected.txt

Tests that Linkifier works correctly.

Debugger was enabled.
listeners added on raw source code: 1
original location: linkifier.html:9
pretty printed location: linkifier.html:12
reverted location: linkifier.html:9
listeners removed from raw source code: 1
Debugger was disabled.
